+++ lib/flutter_2d_amap.dart
@@ -6,7 +6,7 @@ import 'package:flutter/widgets.dart';
 
 export 'src/amap_2d_view.dart';
 export 'src/interface/amap_2d_controller.dart';
-export 'src/poisearch_model.dart';
+export 'src/poi_search_model.dart';
 
 class Flutter2dAMap {
   static const MethodChannel _channel = MethodChannel('plugins.weilu/flutter_2d_amap_');
